requirements.txt
 # Phase 1 Ingestion
 sec-edgar-downloader
python-dotenv

#Utilities
python-dotenv